Extract worker related types and some minor code cleanup
