+++ src/types/worker.d.ts
@@ -1,0 +1,15 @@
+type Event = {
+  cron: string;
+  type: string;
+  scheduledTime: number;
+};
+
+type EnvironmentVariables = {
+  [key: string]: string;
+  KV: KVNamespace;
+};
+
+export {
+    Event,
+    EnvironmentVariables
+};
